+++ modules-js/clickButtonPhysical.js
@@ -1,8 +1,8 @@
 import {keyboardRussian} from "./language-Russian.js";
 import {keyboardEnglish} from "./language-English.js";
-import {changeLanguage} from "./index.js";
-import {changeCapsLock} from "./index.js";
-import {changeShift} from "./index.js";
+import {changeLanguage} from "../index.js";
+import {changeCapsLock} from "../index.js";
+import {changeShift} from "../index.js";
 import {writeSimbol} from "./writeSymbol.js";
 
 let buttonClick;
