+++ modules-js/writeSymbol.js
