+++ style.scss
@@ -1,5 +1,5 @@
 .body {
-  background-image: url("./img_background.jpg");
+  background-image: url("./assets/img_background.jpg");
   background-repeat: no-repeat;
   background-size: cover;
   font-family: 'Montserrat', sans-serif;
